backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
httpx==0.26.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
